pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "makefolio"
version = "0.1.2"
description = "A modern, fast static site generator for professional portfolio websites"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Makefolio Contributors"}
]
keywords = ["static-site-generator", "portfolio", "blog", "markdown"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP :: Site Management",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "click>=8.1.7",
    "jinja2>=3.1.3",
    "markdown>=3.5.1",
    "pyyaml>=6.0.1",
    "watchdog>=3.0.0",
    "python-frontmatter>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.12.0",
    "pytest>=7.4.4",
    "pytest-cov>=4.1.0",
    "bump-my-version>=0.15.0",
]

[project.scripts]
makefolio = "makefolio.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
makefolio = ["themes/**/*"]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

